+++ ReleaseNotes/release_notes.md
@@ -13,7 +13,7 @@
 ## Bug fixes:
 
  1. Fix some tiny incorrectnes in `Ostap/MatrixUtils2.h`
-
+ 1. Fix small problem in `ostap.utils.utils.KeepCWD` context manager 
 
 # v1.6.0.0